# gateway/app.py
import os, sys
sys.path.insert(0, os.path.dirname(__file__))
import yaml
from flask import Flask, render_template, request, redirect, url_for, session, flash, jsonify
from flask_bcrypt import Bcrypt
from datetime import datetime, timedelta
from shared.config import REDES_SOCIALES
from shared.service_client import auth_client, pacientes_client, citas_client, pagos_client, notas_client
from shared.audit import log_accion
from shared.proc import call_proc, call_proc_one, call_proc_execute

bcrypt = Bcrypt()


def _load_swagger():
    path = os.path.join(os.path.dirname(__file__), "..", "swagger.yaml")
    try:
        with open(path, "r", encoding="utf-8") as f:
            return yaml.safe_load(f)
    except Exception:
        return None


def create_app():
    app = Flask(__name__, template_folder="../templates", static_folder="../static")
    app.secret_key = os.getenv("SECRET_KEY", os.urandom(32).hex())
    bcrypt.init_app(app)

    try:
        from flasgger import Swagger
        specs = _load_swagger()
        if specs:
            Swagger(app, template=specs)
        else:
            Swagger(app)
    except Exception:
        pass

    @app.context_processor
    def inject():
        return {
            "REDES": REDES_SOCIALES,
            "current_user": {
                "id": session.get("usuario_id"),
                "nombre": session.get("usuario_nombre"),
                "rol": session.get("rol"),
                "es_admin": session.get("rol") == "admin",
                "es_terapeuta": session.get("rol") == "terapeuta",
            }
        }

    @app.route("/")
    def index():
        return render_template("index.html")

    @app.route("/login", methods=["GET", "POST"])
    def login():
        if request.method == "POST":
            correo = request.form["correo"].strip()
            clave = request.form["clave"].strip()
            if not correo or not clave:
                flash("Completa todos los campos.")
                return redirect(url_for("login"))

            data, status = auth_client.post("/api/auth/login", {"correo": correo, "clave": clave})
            if status == 200 and data.get("success"):
                u = data["usuario"]
                session["usuario_id"] = u["id"]
                session["usuario_nombre"] = u["nombre"]
                session["rol"] = u["rol"]
                session["correo"] = u["correo"]
                if u["rol"] == "admin":
                    return redirect(url_for("panel_admin"))
                return redirect(url_for("interfaz"))
            else:
                flash(data.get("error", "Credenciales incorrectas."))

        return render_template("login.html")

    @app.route("/logout")
    def logout():
        session.clear()
        return redirect(url_for("login"))

    @app.route("/interfaz")
    def interfaz():
        if "usuario_id" not in session:
            return redirect(url_for("login"))

        fecha_str = request.args.get("fecha", datetime.today().strftime("%Y-%m-%d"))
        try:
            fecha_obj = datetime.strptime(fecha_str, "%Y-%m-%d")
        except ValueError:
            fecha_obj = datetime.today()

        ayer = (fecha_obj - timedelta(days=1)).strftime("%Y-%m-%d")
        manana = (fecha_obj + timedelta(days=1)).strftime("%Y-%m-%d")
        dias = ["Domingo", "Lunes", "Martes", "Miercoles", "Jueves", "Viernes", "Sabado"]
        dia_semana = dias[fecha_obj.weekday() + 1 if fecha_obj.weekday() < 6 else 0]
        fecha_display = fecha_obj.strftime("%d/%m/%Y")
        es_hoy = fecha_str == datetime.today().strftime("%Y-%m-%d")
        es_admin = session.get("rol") == "admin"

        params = f"?fecha={fecha_str}&estado=programada"
        data, _ = citas_client.get(f"/api/citas{params}")
        pacientes = data.get("citas", [])

        return render_template(
            "interfaz.html", pacientes=pacientes,
            fecha_actual=fecha_str, fecha_display=fecha_display,
            dia_semana=dia_semana, ayer=ayer, manana=manana,
            es_hoy=es_hoy, nombre_usuario=session.get("usuario_nombre"), es_admin=es_admin,
        )

    @app.route("/guardar_descripcion", methods=["POST"])
    def guardar_descripcion():
        if "usuario_id" not in session:
            return redirect(url_for("login"))
        historial_id = request.form.get("historial_id")
        descripcion = request.form.get("descripcion", "").strip()
        fecha = request.form.get("fecha", datetime.today().strftime("%Y-%m-%d"))

        call_proc_execute("sp_completar_cita", (historial_id, descripcion))
        log_accion(
            usuario_id=session.get("usuario_id"),
            accion="completar_cita",
            tabla_afectada="historial_citas",
            registro_id=historial_id,
            detalle="Cita marcada como completada",
        )
        return redirect(url_for("interfaz", fecha=fecha))

    @app.route("/citas", methods=["GET", "POST"])
    def citas_page():
        data, _ = citas_client.get("/api/citas/terapeutas")
        terapeutas = data.get("terapeutas", [])
        sdata, _ = pacientes_client.get("/api/servicios")
        servicios = sdata.get("servicios", [])

        if request.method == "POST":
            form_data = {k: request.form.get(k, "").strip() for k in
                         ["nombre", "apellido", "dni", "telefono", "medico_id", "fecha_cita", "metodo_pago", "servicio_id"]}
            if not all(form_data.values()):
                flash("campos_vacios")
                return redirect(url_for("citas_page"))

            result, status = citas_client.post("/api/citas", form_data)
            if status == 201 and result.get("success"):
                return redirect(url_for("pago_page", cita_id=result["cita_id"]))
            else:
                flash(result.get("error", "Error al crear cita"))

        return render_template("citas.html", terapeutas=terapeutas, servicios=servicios)

    @app.route("/citas/modificar", methods=["GET", "POST"])
    def modificar_cita_page():
        data, _ = citas_client.get("/api/citas/terapeutas")
        terapeutas = data.get("terapeutas", [])
        citas_encontradas = None
        paso = None
        dni_val = ""
        tel_mask = ""
        error_msg = None

        if request.method == "POST":
            accion = request.form.get("accion")
            if accion == "solicitar":
                dni = request.form.get("dni", "").strip()
                result, status = citas_client.post("/api/citas/otp/solicitar", {"dni": dni, "accion": "modificar"})
                if status == 200 and result.get("success"):
                    paso, dni_val, tel_mask = "verificar", dni, result["tel_mask"]
                else:
                    error_msg = result.get("error", "Error.")
            elif accion == "verificar":
                dni = request.form.get("dni", "").strip()
                otp = request.form.get("otp", "").strip()
                result, _ = citas_client.post("/api/citas/otp/verificar", {"dni": dni, "otp": otp, "accion": "modificar"})
                r = result.get("resultado", "")
                if r == "ok":
                    cdata, _ = citas_client.get(f"/api/citas?dni={dni}&estado=programada")
                    citas_encontradas = cdata.get("citas", [])
                else:
                    mensajes = {"expirado": "El codigo expiro.", "agotado": "Intentos agotados.", "no_existe": "Solicita uno nuevo."}
                    error_msg = mensajes.get(r, f"Codigo incorrecto. Quedan {result.get('restantes', '?')} intento(s).")
                    paso, dni_val = "verificar", dni
            elif accion == "guardar":
                cita_id = request.form.get("cita_id")
                nueva_fecha = request.form.get("fecha_cita", "").strip()
                nuevo_medico = request.form.get("medico_id", "").strip()
                result, status = citas_client.put(f"/api/citas/{cita_id}", {"fecha_cita": nueva_fecha, "medico_id": nuevo_medico})
                if status == 200:
                    flash("exito:Cita modificada correctamente.")
                else:
                    flash(result.get("error", "Error al modificar."))

        return render_template("modificar_cita.html", terapeutas=terapeutas, citas=citas_encontradas,
                               paso=paso, dni=dni_val, tel_mask=tel_mask, error=error_msg)

    @app.route("/citas/cancelar", methods=["GET", "POST"])
    def cancelar_cita_page():
        citas_encontradas = None
        paso = None
        dni_val = ""
        tel_mask = ""
        error_msg = None

        if request.method == "POST":
            accion = request.form.get("accion")
            if accion == "solicitar":
                dni = request.form.get("dni", "").strip()
                result, status = citas_client.post("/api/citas/otp/solicitar", {"dni": dni, "accion": "cancelar"})
                if status == 200 and result.get("success"):
                    paso, dni_val, tel_mask = "verificar", dni, result["tel_mask"]
                else:
                    error_msg = result.get("error", "Error.")
            elif accion == "verificar":
                dni = request.form.get("dni", "").strip()
                otp = request.form.get("otp", "").strip()
                result, _ = citas_client.post("/api/citas/otp/verificar", {"dni": dni, "otp": otp, "accion": "cancelar"})
                r = result.get("resultado", "")
                if r == "ok":
                    cdata, _ = citas_client.get(f"/api/citas?dni={dni}&estado=programada")
                    citas_encontradas = cdata.get("citas", [])
                else:
                    mensajes = {"expirado": "El codigo expiro.", "agotado": "Intentos agotados.", "no_existe": "Solicita uno nuevo."}
                    error_msg = mensajes.get(r, f"Codigo incorrecto. Quedan {result.get('restantes', '?')} intento(s).")
                    paso, dni_val = "verificar", dni
            elif accion == "confirmar":
                cita_id = request.form.get("cita_id")
                result, _ = citas_client.delete(f"/api/citas/{cita_id}")
                flash("exito:Tu cita ha sido cancelada correctamente.")

        return render_template("cancelar_cita.html", citas=citas_encontradas,
                               paso=paso, dni=dni_val, tel_mask=tel_mask, error=error_msg)

    @app.route("/pago")
    def pago_page():
        cita_id = request.args.get("cita_id")
        data, _ = pagos_client.get(f"/api/pagos/{cita_id}")
        pago = data.get("pago", {})

        cdata, _ = citas_client.get(f"/api/citas/{cita_id}")
        cita = cdata.get("cita", {})

        if pago.get("estado_pago") == "pagado":
            return redirect(url_for("retorno_page", cita_id=cita_id))

        cita["monto"] = pago.get("monto", 0)
        cita["metodo_pago"] = pago.get("metodo_pago", "")
        cita["estado_pago"] = pago.get("estado_pago", "pendiente")

        from services.pagos_service.providers import NiubizClient
        niubiz = NiubizClient()
        return render_template("pago.html", cita=cita, niubiz_sdk_url=niubiz.sdk_url, niubiz_mode=niubiz.mode)

    @app.route("/retorno")
    def retorno_page():
        cita_id = request.args.get("cita_id")
        data, _ = pagos_client.get(f"/api/pagos/{cita_id}")
        pago = data.get("pago", {})

        if pago.get("estado_pago") != "pagado":
            return redirect(url_for("pago_page", cita_id=cita_id))

        cdata, _ = citas_client.get(f"/api/citas/{cita_id}")
        cita = cdata.get("cita", {})
        cita["monto"] = pago.get("monto", 0)
        cita["metodo_pago"] = pago.get("metodo_pago", "")
        cita["referencia"] = pago.get("referencia")
        cita["transaccion_id"] = pago.get("transaccion_id")

        fecha_fmt = cita.get("fecha_cita", "")
        try:
            fecha_fmt = datetime.strptime(str(cita["fecha_cita"]), "%Y-%m-%d").strftime("%d/%m/%Y")
        except Exception:
            pass

        return render_template("retorno.html", cita_id=cita_id, cita=cita, fecha_fmt=fecha_fmt)

    @app.route("/panel_admin", methods=["GET", "POST"])
    def panel_admin():
        if "usuario_id" not in session or session.get("rol") != "admin":
            return redirect(url_for("login"))

        if request.method == "POST":
            accion = request.form.get("accion")
            if accion == "registrar":
                nombre = request.form.get("nombre", "").strip()
                especialidad = request.form.get("especialidad", "").strip()
                correo = request.form.get("correo", "").strip()
                clave = request.form.get("clave", "").strip()
                precio = request.form.get("precio", "").strip()

                if nombre and especialidad and correo and clave:
                    if call_proc_one("sp_obtener_usuario_por_nombre", (nombre,)):
                        flash("Ya existe un terapeuta con ese nombre.")
                    else:
                        try:
                            precio_num = float(precio) if precio else None
                        except ValueError:
                            precio_num = None
                        telefono = request.form.get("telefono", "").strip()
                        clave_hash = bcrypt.generate_password_hash(clave).decode("utf-8")
                        rol = call_proc_one("sp_obtener_rol_id_terapeuta")
                        rol_id = rol["id"] if rol else 2
                        creado = call_proc_one("sp_crear_usuario_admin", (
                            nombre, correo, telefono or None, clave_hash, rol_id,
                        ))
                        usuario_id = creado["id"] if creado else None
                        esp = call_proc_one("sp_obtener_especialidad_id", (especialidad,))
                        esp_id = esp["id"] if esp else None
                        call_proc_execute("sp_crear_terapeuta", (usuario_id, esp_id, precio_num))
                        log_accion(
                            usuario_id=session.get("usuario_id"),
                            accion="crear_usuario",
                            tabla_afectada="usuarios",
                            registro_id=usuario_id,
                            detalle=f"Terapeuta creado: nombre={nombre}, correo={correo}",
                        )
                        flash("exito:Terapeuta registrado correctamente.")
                else:
                    flash("Completa todos los campos.")

            elif accion == "actualizar_precio":
                mid = request.form.get("medico_id")
                precio = request.form.get("precio", "").strip()
                if mid and precio:
                    try:
                        call_proc_execute("sp_actualizar_precio", (mid, float(precio)))
                        flash("exito:Precio actualizado.")
                    except ValueError:
                        flash("Precio invalido.")

            elif accion == "eliminar":
                mid = request.form.get("medico_id")
                if mid:
                    terapeuta = call_proc_one("sp_obtener_usuario_id_terapeuta", (mid,))
                    if terapeuta:
                        call_proc_execute("sp_set_terapeuta_activo", (mid, 0))
                        if terapeuta.get("usuario_id"):
                            call_proc_execute("sp_set_usuario_activo", (terapeuta["usuario_id"], 0))
                        log_accion(
                            usuario_id=session.get("usuario_id"),
                            accion="desactivar_usuario",
                            tabla_afectada="usuarios",
                            registro_id=terapeuta.get("usuario_id"),
                            detalle=f"Terapeuta id={mid} desactivado",
                        )
                        flash("exito:Terapeuta desactivado.")

            elif accion == "reactivar":
                mid = request.form.get("medico_id")
                if mid:
                    terapeuta = call_proc_one("sp_obtener_usuario_id_terapeuta", (mid,))
                    if terapeuta:
                        call_proc_execute("sp_set_terapeuta_activo", (mid, 1))
                        if terapeuta.get("usuario_id"):
                            call_proc_execute("sp_set_usuario_activo", (terapeuta["usuario_id"], 1))
                        flash("exito:Terapeuta reactivado.")

            elif accion == "cambiar_clave":
                mid = request.form.get("medico_id")
                nc = request.form.get("nueva_clave", "").strip()
                if mid and nc:
                    terapeuta = call_proc_one("sp_obtener_usuario_id_terapeuta", (mid,))
                    if terapeuta and terapeuta.get("usuario_id"):
                        h = bcrypt.generate_password_hash(nc).decode("utf-8")
                        call_proc_execute("sp_cambiar_clave_usuario", (terapeuta["usuario_id"], h))
                        flash("exito:Contrasena actualizada.")

            return redirect(url_for("panel_admin"))

        medicos = call_proc("sp_consultar_medicos_admin")
        especialidades = call_proc("sp_listar_especialidades")
        usuarios = call_proc("sp_listar_usuarios")

        params = "?fecha=" + datetime.today().strftime("%Y-%m-%d") + "&estado=programada"
        cdata, _ = citas_client.get(f"/api/citas{params}")
        proximas_citas = cdata.get("citas", [])

        return render_template("paneladmin.html", medicos=medicos, proximas_citas=proximas_citas,
                               especialidades=especialidades, usuarios=usuarios)

    @app.route("/panel_admin/cancelar_cita/<int:cita_id>", methods=["POST"])
    def admin_cancelar_cita(cita_id):
        if "usuario_id" not in session or session.get("rol") != "admin":
            return redirect(url_for("login"))
        citas_client.delete(f"/api/citas/{cita_id}")
        flash("exito:Cita cancelada.")
        return redirect(url_for("panel_admin"))

    @app.route("/pacientes")
    def pacientes_page():
        if "usuario_id" not in session or session.get("rol") != "admin":
            return redirect(url_for("login"))
        data, _ = pacientes_client.get("/api/pacientes")
        return render_template("pacientes.html", pacientes=data.get("pacientes", []))

    @app.route("/pacientes/<paciente_dni>")
    def detalle_paciente_page(paciente_dni):
        if "usuario_id" not in session or session.get("rol") != "admin":
            return redirect(url_for("login"))
        data, _ = pacientes_client.get(f"/api/pacientes/{paciente_dni}")
        paciente = data.get("paciente", {})
        historial = data.get("historial", [])
        paciente_id_val = paciente.get("id")

        paquetes, evaluaciones, consentimientos = [], [], []
        if paciente_id_val:
            pdata, _ = pacientes_client.get(f"/api/pacientes/{paciente_id_val}/paquetes")
            paquetes = pdata.get("paquetes", [])
            edata, _ = pacientes_client.get(f"/api/pacientes/{paciente_id_val}/evaluaciones")
            evaluaciones = edata.get("evaluaciones", [])
            cdata, _ = pacientes_client.get(f"/api/pacientes/{paciente_id_val}/consentimientos")
            consentimientos = cdata.get("consentimientos", [])

        return render_template("detalle_paciente.html", paciente=paciente,
                               historial=historial, paquetes=paquetes,
                               evaluaciones=evaluaciones, consentimientos=consentimientos)

    @app.route("/notas/<int:cita_id>", methods=["GET", "POST"])
    def notas_page(cita_id):
        if "usuario_id" not in session:
            return redirect(url_for("login"))

        cdata, _ = citas_client.get(f"/api/citas/{cita_id}")
        cita = cdata.get("cita", {})

        if request.method == "POST":
            nota = request.form.get("nota", "").strip()
            diagnostico = request.form.get("diagnostico", "").strip()
            notas_client.post(f"/api/notas/{cita_id}", {
                "nota": nota, "diagnostico": diagnostico,
                "terapeuta_id": None,
                "paciente_id": cita.get("paciente_id"),
            })
            flash("exito:Nota clinica guardada.")
            return redirect(url_for("notas_page", cita_id=cita_id))

        ndata, _ = notas_client.get(f"/api/notas/{cita_id}")
        return render_template("notas_cita.html", cita=cita, notas=ndata.get("notas", []))

    @app.route("/api/verificar_dni", methods=["POST"])
    def verificar_dni():
        dni = (request.json or {}).get("dni", "").strip()
        data, status = pacientes_client.post("/api/pacientes/buscar_dni", {"dni": dni})
        return jsonify(data), status

    @app.route("/api/estadisticas")
    def api_estadisticas():
        data, _ = citas_client.get("/api/citas/estadisticas")
        return jsonify(data)

    @app.route("/api/admin/metricas")
    def api_admin_metricas():
        if "usuario_id" not in session or session.get("rol") != "admin":
            return jsonify({"error": "No autorizado"}), 401

        filas = call_proc("sp_dashboard_metricas_admin")

        citas_estado = {"programada": 0, "completada": 0, "cancelada": 0}
        ingresos_metodo = {}
        ingresos_mes = {}
        sesiones = {"usadas": 0, "restantes": 0}

        for f in filas:
            tipo = f.get("tipo")
            etiqueta = f.get("etiqueta")
            valor = float(f.get("valor") or 0)
            periodo = f.get("periodo")

            if tipo == "citas_estado" and etiqueta in citas_estado:
                citas_estado[etiqueta] = int(valor)
            elif tipo == "ingresos_metodo":
                ingresos_metodo[etiqueta] = valor
            elif tipo == "ingresos_mes":
                ingresos_mes.setdefault(periodo, {})[etiqueta] = valor
            elif tipo == "sesiones" and etiqueta in sesiones:
                sesiones[etiqueta] = int(valor)

        meses = sorted(ingresos_mes.keys())
        metodos = sorted(set(ingresos_metodo.keys()) | {
            m for datos in ingresos_mes.values() for m in datos.keys()
        })

        return jsonify({
            "success": True,
            "citas_por_estado": citas_estado,
            "ingresos_por_metodo": ingresos_metodo,
            "ingresos_por_mes": {
                "meses": meses,
                "metodos": metodos,
                "series": ingresos_mes,
            },
            "sesiones": sesiones,
        })

    return app