# swagger.yaml
...
swagger: "2.0"
info:
  title: "MOOVA Clinic API"
  description: |
    API del sistema de clinica de rehabilitacion MOOVA CLINIC.
    Backend distribuido: microservicios auth, pacientes, citas,
    pagos y notas, mas el gateway central. Esta especificacion
    documenta todos los endpoints accesibles.
  version: "1.0.0"
basePath: "/"
schemes:
  - http
  - https
produces:
  - application/json
consumes:
  - application/json
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Inicio de sesion
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              correo: { type: string, example: "admin@moova.clinic" }
              clave: { type: string, example: "123456" }
      responses:
        "200":
          description: Login correcto
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              usuario:
                type: object
                properties:
                  id: { type: integer }
                  nombre: { type: string }
                  correo: { type: string }
                  rol: { type: string }
        "400": { description: Campos incompletos }
        "401": { description: Credenciales incorrectas }
...
  /api/pacientes:
    get:
      tags: [pacientes]
      summary: Listar pacientes con total y ultima cita
      responses:
        "200":
          description: Lista de pacientes
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              pacientes:
                type: array
                items:
                  type: object
    post:
      tags: [pacientes]
      summary: Crear paciente
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              nombre: { type: string, example: "Juan" }
              apellido: { type: string, example: "Perez" }
              dni: { type: string, example: "12345678" }
              telefono: { type: string, example: "999888777" }
              email: { type: string }
              fecha_nacimiento: { type: string, format: date }
              sexo: { type: string, enum: ["M", "F", "otro"] }
              direccion: { type: string }
              seguro: { type: string }
      responses:
        "201":
          description: Paciente creado
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              dni: { type: string }
        "400": { description: Campos requeridos }
        "409": { description: DNI ya existe }
  /api/pacientes/{dni}:
    get:
      tags: [pacientes]
      summary: Detalle de paciente con historial
      parameters:
        - name: dni
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Paciente y su historial
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              paciente: { type: object }
              historial: { type: array }
        "404": { description: Paciente no encontrado }
    put:
      tags: [pacientes]
      summary: Actualizar paciente
      parameters:
        - name: dni
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              nombre: { type: string }
              apellido: { type: string }
              telefono: { type: string }
              email: { type: string }
              estado: { type: string }
              fecha_nacimiento: { type: string, format: date }
              sexo: { type: string }
              direccion: { type: string }
              seguro: { type: string }
      responses:
        "200":
          description: Paciente actualizado
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
        "400": { description: Nada que actualizar }
  /api/pacientes/buscar_dni:
    post:
      tags: [pacientes]
      summary: Buscar datos de DNI en API externa (RENIEC proxy)
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              dni: { type: string, example: "12345678" }
      responses:
        "200":
          description: Datos encontrados
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: { type: object }
        "400": { description: DNI invalido }
        "404": { description: DNI no encontrado }
...
  /api/pacientes/{paciente_id}/paquetes:
    get:
      tags: [pacientes]
      summary: Listar paquetes de sesiones de un paciente
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
      responses:
        "200":
          description: Paquetes del paciente
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              paquetes: { type: array }
    post:
      tags: [pacientes]
      summary: Crear paquete de sesiones
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              servicio_id: { type: integer }
              total_sesiones: { type: integer }
              fecha_compra: { type: string, format: date }
              fecha_vencimiento: { type: string, format: date }
      responses:
        "201":
          description: Paquete creado
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              paquete_id: { type: integer }
        "400": { description: Campos requeridos }
  /api/pacientes/{paciente_id}/evaluaciones:
    get:
      tags: [pacientes]
      summary: Listar evaluaciones iniciales de un paciente
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
      responses:
        "200":
          description: Evaluaciones
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              evaluaciones: { type: array }
    post:
      tags: [pacientes]
      summary: Crear evaluacion inicial
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              terapeuta_id: { type: integer }
              motivo_consulta: { type: string }
              escala_dolor_eva: { type: integer, minimum: 0, maximum: 10 }
              rango_movimiento: { type: string }
              objetivos_terapeuticos: { type: string }
      responses:
        "201":
          description: Evaluacion creada
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              evaluacion_id: { type: integer }
        "400": { description: Campos requeridos }
  /api/pacientes/{paciente_id}/consentimientos:
    get:
      tags: [pacientes]
      summary: Listar consentimientos de un paciente
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
      responses:
        "200":
          description: Consentimientos
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              consentimientos: { type: array }
    post:
      tags: [pacientes]
      summary: Crear consentimiento
      parameters:
        - name: paciente_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              tipo: { type: string, example: "tratamiento" }
              texto_version: { type: string, example: "v1.0" }
      responses:
        "201":
          description: Consentimiento creado
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              consentimiento_id: { type: integer }
        "400": { description: Campos requeridos }
...
  /api/citas:
    get:
      tags: [citas]
      summary: Listar citas con filtros opcionales
      parameters:
        - name: dni
          in: query
          type: string
        - name: fecha
          in: query
          type: string
          format: date
        - name: estado
          in: query
          type: string
          default: programada
        - name: medico_id
          in: query
          type: integer
      responses:
        "200":
          description: Lista de citas
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              total: { type: integer }
              citas: { type: array }
    post:
      tags: [citas]
      summary: Crear cita (y paciente si no existe, y anticipo)
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              nombre: { type: string }
              apellido: { type: string }
              dni: { type: string }
              telefono: { type: string }
              medico_id: { type: integer }
              fecha_cita: { type: string, format: date }
              servicio_id: { type: integer }
              metodo_pago: { type: string, enum: ["yape", "plin", "tarjeta", "transferencia"] }
      responses:
        "201":
          description: Cita creada con costo y anticipo
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              cita_id: { type: integer }
              costo: { type: number }
              anticipo: { type: number }
        "400": { description: Campo requerido / fecha invalida }
        "404": { description: Medico no encontrado }
        "409": { description: Medico ya tiene cita ese dia }
...
  /api/citas/otp/solicitar:
    post:
      tags: [citas]
      summary: Solicitar codigo OTP por SMS para modificar/cancelar
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              dni: { type: string, example: "12345678" }
              accion: { type: string, enum: ["modificar", "cancelar"], default: "modificar" }
      responses:
        "200":
          description: OTP enviado
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              tel_mask: { type: string, example: "999***8777" }
        "404": { description: No hay citas programadas para ese DNI }
        "500": { description: No se pudo enviar el SMS }
  /api/citas/otp/verificar:
    post:
      tags: [citas]
      summary: Verificar codigo OTP
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              dni: { type: string }
              otp: { type: string }
              accion: { type: string, enum: ["modificar", "cancelar"], default: "modificar" }
      responses:
        "200":
          description: Resultado de verificacion
          schema:
            type: object
            properties:
              resultado: { type: string, enum: ["ok", "incorrecto", "expirado", "agotado", "no_existe"] }
              restantes: { type: integer }
...